requirements.txt
discord.py
python-dotenv
praw
